solar-seas/python/requirements.txt
strawberry-graphql
fastapi
uvicorn